third_party/blink/web_tests/external/wpt/css/css-grid/subgrid/grid-template-invalid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['grid-template-rows'] = "subgrid repeat(auto-fit, [a])" should not set the property value
  assert_equals: expected "" but got "subgrid repeat(auto-fit, [a])"
[FAIL] e.style['grid-template-columns'] = "subgrid repeat(auto-fit, [a])" should not set the property value
  assert_equals: expected "" but got "subgrid repeat(auto-fit, [a])"
Harness: the test ran to completion.
